Moved plugins into own files and updated imports
