+++ src/Transpiler.js
@@ -83,266 +83,3 @@ export class TranspilerError extends Error {
         super(message);
     }
 }
-
-export class ContentTransformerPlugin {
-    _name;
-    _parameters;
-    _preProcess;
-
-    constructor(name = "", parameters = [], preProcess = false) {
-        this._name = name;
-        this._parameters = parameters;
-        this._preProcess = preProcess;
-    }
-
-    get name() { return this._name; }
-    set name(name) { this._name = name; }
-    get parameters() { return this._parameters; }
-    set parameters(parameters) { this._parameters = parameters; }
-
-    isPreProcessor() { return this._preProcess; }
-
-    /**
-     * Parses the given block, returning a string.
-     * 
-     * May have side effects.
-     * @param {*} block 
-     */
-    transform(block) {
-        throw UnimplementedError("Not implemented: Plugin unable to parse @ block");
-    }
-}
-
-export class PreProcessingContentPlugin extends ContentTransformerPlugin {
-    constructor(name = "", parameters = []) {
-        super(name, parameters, true);
-    }
-
-    // Don't return anything from `transform()` if the pre-processor removes
-    // a block, otherwise return the pre-processed block
-}
-
-export class TemplatePreProcessor extends PreProcessingContentPlugin {
-    _weaver;
-
-    /**
-     * 
-     * @param {WeaveTemplatePlugin} weaver 
-     */
-    constructor(weaver) {
-        super("template", ["name", "args", "type"]);
-        this._weaver = weaver;
-    }
-
-    transform(block) {
-        this._weaver.addTemplate(block);
-        return "";
-    }
-}
-
-export class WeaveTemplatePlugin extends ContentTransformerPlugin {
-    _templates;
-
-    constructor() {
-        super("weave", ["name", "argsType"]);
-        this._templates = {};
-    }
-
-    transform(block) {
-        const weaveNameParam = block.parameters.find(param => param.name === "name");
-        if (!weaveNameParam) throw new Error("Template weave parameter 'name' not supplied");
-        const template = this._templates[weaveNameParam.value];
-        if (!template) throw new Error("Could not find template with name '" + weaveNameParam.value + "'");
-        if (block.type === BlockType.INLINE_PLUGIN) return this.weaveInline(block, template);
-        const weaveArgs = JSON.parse(`{${block.value.value}}`);
-        return this.weave(template, weaveArgs);
-    }
-
-    addTemplate(block) {
-        const templateNameParameter = block.parameters.find(param => param.name === "name");
-        if (!templateNameParameter) throw new Error("Cannot create nameless template");
-        this._templates[templateNameParameter.value] = block;
-    }
-
-    weave(templateBlock, weaveArgs) {
-        // expected params of template
-        let expectedParams = templateBlock.parameters.find(param => param.name === "args");
-        if (expectedParams) expectedParams = expectedParams.length === 0 ? [] : expectedParams.value.split(" ");
-        
-        // prepare output
-        let output = templateBlock.value.value;
-        if (expectedParams) expectedParams.forEach(param => output = output.replace(`@${param};`, weaveArgs[param] || ""));
-
-        return output;
-    }
-
-    weaveInline(block, template) {
-        // capture reserved parameter 'name', every other param is a weave arg
-        const weaveArgs = {};
-        const params = block.parameters.filter(param => param.name !== "name");
-        for (const param of params) {
-            weaveArgs[param.name] = param.value;
-        }
-        return this.weave(template, weaveArgs);
-    }
-}
-
-export class DocumentTransformerPlugin {
-    /**
-     * Transforms the given content into a document such as React component or HTML document.
-     * @param {*} content 
-     * @param {*} args 
-     */
-     transform(src, args) {
-        throw Error("Not implemented: Plugin unable to transform document");
-    }
-}
-
-export class HtmlDocumentTransformer extends DocumentTransformerPlugin {
-    _preamble;
-    _postamble;
-
-    constructor(preamble = "", postamble = "") {
-        super();
-        this._preamble = preamble;
-        this._postamble = postamble;
-    }
-
-    transform(src, args) {
-        const argMap = this._validateArgs(args);
-        return `<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    ${argMap.title}
-    ${argMap.links.join("\n" + " ".repeat(4)) || ""}
-</head>
-<body>
-    ${this._preamble}
-    ${src}
-    ${this._postamble}
-    ${argMap.scripts.join("\n" + " ".repeat(4)) || ""}
-</body>
-</html>
-`;
-    }
-
-    _validateArgs(args) {
-        const argMap = {
-            links: [],
-            scripts: []
-        };
-
-        // page title
-        if (!args.title) {
-            console.warn("WARNING: No <title> provided");
-            argMap.title = "<title>Placeholder</title>";
-        } else argMap.title = `<title>${args.title}</title>`;
-
-        // page links
-        if (!args.links) console.warn("WARNING: No links provided, embedding and inline CSS can be messy...");
-        else argMap.links = args.links.map(arg => `<link rel="stylesheet" href="${arg}" />`);
-
-        if (!args.scripts) console.warn("WARNING: No scripts provided, embedded and inline JS can be messy...");
-        else argMap.scripts = args.scripts.map(arg => `<script src="${arg}"></script>`);
-
-        return argMap;
-    }
-}
-
-
-export class JSTransformer extends ContentTransformerPlugin {
-
-    _context = {};
-
-    constructor() {
-        super("js")
-    }
-
-    getArgs(parameters) {
-        const args = {};
-        parameters.forEach(parameter => args[parameter.name] = parameter);
-        return args;
-    }
-
-    transform(block) {
-        const args = this.getArgs(block.parameters);
-        if (block.type === BlockType.INLINE_PLUGIN) return this.transformInline(args);
-        return this.transformBlock(block, args);
-    }
-
-    transformBlock(block, args) {
-        if (args.call) {
-            const tryCallResult = this.tryCall(args);
-            if (tryCallResult) return tryCallResult;
-        }
-
-        if (args.name && args.defer && args.defer.value === "true") {
-            this._context[args.name.value] = new Function("context", block.value.value);
-            return "";
-        } else if (args.name) {
-            // if not deferring execution, store a copy of the function and execute
-            this._context[args.name.value] = new Function("context", block.value.value);
-        }
-        return Function("context", block.value.value)(this._context) || "";
-    }
-
-    transformInline(args) {
-        this.checkForRequiredInlineArgs(args);
-        const tryCallResult = this.tryCall(args)
-        if (tryCallResult) return tryCallResult;
-
-        if (args.name && args.defer && args.defer.value === "true") {
-            this._context[args.name.value] = new Function("context", args.value.value);
-            return "";
-        } else if (args.name) {
-            // if not deferring execution, store a copy of the function and execute
-            this._context[args.name.value] = new Function("context", args.value.value);
-        }
-        return Function("context", args.value.value)(this._context) || "";
-    }
-
-    checkForRequiredInlineArgs(args) {
-        if (!args.call && !args.value) throw new TranspilerError("JSTransformer: Could not find 'value' or 'call' argument for inline plugin.");
-        if (args.call && args.value) throw new TranspilerError("JSTransformer: 'call' and 'value' cannot both be specified.");
-    }
-
-    tryCall(args) {
-        if (args.call) try {
-            return this._context[args.call.value](this._context);
-        } catch (error) {
-            console.warn("Function with name '" + args.call.value + "' does not exist in the current context.");
-        }
-        return null;
-    }
-}
-
-export class LiteralTransformer extends ContentTransformerPlugin {
-    constructor() {
-        super("lit")
-    }
-
-    transform(block) {
-        return block.value.value;
-    }
-}
-
-/**
-* Internal plugin for passing arguments from a document to a DocTypePlugin's transform method.
-*/
-export class DocumentArgumentsTransformer extends ContentTransformerPlugin {
-    constructor() {
-        super("docArgs");
-    }
- 
-    transform(block) {
-        try {
-            return JSON.parse("{" + block.value.value + "}");
-        } catch (e) {
-            console.warn(e);
-            return {};
-        }
-    }
-}
