+++ src/plugins/DocumentTransformerPlugin.js
@@ -1,0 +1,10 @@
+export default class DocumentTransformerPlugin {
+    /**
+     * Transforms the given content into a document such as React component or HTML document.
+     * @param {*} content 
+     * @param {*} args 
+     */
+     transform(src, args) {
+        throw Error("Not implemented: Plugin unable to transform document");
+    }
+}
