+++ src/plugins/built_in/LiteralTransformer.js
@@ -1,0 +1,11 @@
+import ContentTransformerPlugin from "../ContentTransformerPlugin.js";
+
+export default class LiteralTransformer extends ContentTransformerPlugin {
+    constructor() {
+        super("lit")
+    }
+
+    transform(block) {
+        return block.value.value;
+    }
+}
